Switch fetch severity from ERR -> INFO.
This patch addresses a pet peeve of mine. During cleanup on a failed
robotest run, the "Fetching node logs" logging statements show up as
errors when in reality they're running to completion as expected.

I refactored logic surrounding the handling of apiserver logging to
clarify intent.  The special treatment of the first item of a was
confusing without the additional context that the first item was the
kubernetes api-server. Bonus: this refactoring saves a function or two,
duplicated logic, and ~10 SLOC.