Make required `recover` parameters optional.
When boolean parameters are 'required', they are actually 'required to
be true' because go-playground/validator cannot distinguish between a
struct's default value and a user provided one (unless it is e.g. a
pointer to a value).  In 2020, there is little reason that
`expand_before_shrink` or `pwroff_before_remove` should be required.
These are knobs that manual testers may care about, but automation
should use the defaults for.