Conditionally push an extra quay.io image tag.
Previously, `make publish` would always tag images with the version
and another tag, 'latest' if unspecified by the user.

In an effort to reduce the use of floating tags and encourage pinned use
of semantic versions, the build will now skip publishing a tag unless it
is specified by user/automation calling `make publish`.

This helps us move the TAG parameter in jenkins from a choice that must
have a default to a string that may be empty.

Contributes to #200.