Add '-retries' cli flag.
While debugging, I frequently want to run a test once, with no option
for robotest to re-attempt it if/when it fails.  When robotest auto
retries, I must then abort the program and clean up the resources from
the next attempt. Previously, to do this I needed to edit
lib/defaults/defaults.go.