Change apiserverNode to allocate new arrays instead of subslices.
Formerly calling apiserverNode sorted the slice that was passed to it,
and then passed back subslices pointing to the same array. Sharing the
array like this made the code prone to issues if the elements of the
slice were ever altered later and would unexpectedly alter the ordering
of the input array.

I burned a good hour debugging an issue related to this when a subcall
to apiserverNode unexpectedly reorered some elements of a slice I was
working with.