Refactor expand and shrink codepaths.
In preparation for shrink work (#188), this patch refactors Expand and
the new Shrink to have consistent signatures.  It also adds JoinNode and
retouches RemoveNode to be consistent & independently useful singular
pieces of code.

The main motivation for this is the fact that shrink is not
parallelizable:

  https://github.com/gravitational/robotest/pull/229#discussion_r428221568

Expand/Join was already serializied, and the signature reworked for
consistency.