Refactor terraform output logging.
In logrus 1.0.3 all log fields started being escaped.  This multiline
terraform stdout was collapsed into a single line, and became quite
unreadable.

To counteract this, put the stdout into the log message, instead of one
of the fields.  A bit of a hack, but it keeps console logs a bit more
readable.

Remove a log statement (now unnecessary, with all stdout captured) and
decrease the loglevel to debug of terraform command logs (as they're now
a deal more verbose.

Fixes #244.