Use python3 when python is not available.
SUSE 15 does not distribute python in the default image, but it does
distribute python3.  SUSE 12 sp5 has both python and python3 in the
image, but its python3 is 3.4, which is under the required 3.5 for the
awscli package.  Thus to install awscli, we determine which python is
available, with a preference for plain ol' python.