Switch from `suse` to `sles` distro naming.
This is for clarity. Robotest doesn't presently offer openSUSE as a
target distro, but it does offer SLES. `suse` aliases are retained for
backwards compatibility, but may eventually be removed or point to
openSUSE images.

Furthermore, sles requires specifying the service pack -- to accommodate
that we may want to test on different service packs.